api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.15
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.10.0
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
ordered-set==4.1.0
packaging==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
rsa==4.9
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
